Download from MtGox the value of Bitcoins (Euro and USD)

+++ src/org/silix/the9ull/microbit/control/GetInfoBeanRemote.java
@@ -1,5 +1,7 @@
 package org.silix.the9ull.microbit.control;
 
+import java.math.BigDecimal;
+
 import javax.ejb.Remote;
 
 @Remote
@@ -7,8 +9,8 @@ public interface GetInfoBeanRemote {
 	
 	public long numberOfUsers();
 	
-    public double valueBtcUsd();
-    public double valueBtcEur();
+    public BigDecimal valueBtcUsd();
+    public BigDecimal valueBtcEur();
     
     public int getIdFromAddress(String address);
     public String getAddressFromId(int id);
